Move `CompiledJarFile` and `JarEntries` to `codegen` package
